Eliminate Jasper compilation warnings in WIN64 build.

+++ jp2/src/appl/jasper.c
@@ -447,12 +447,12 @@ void cmdopts_destroy(cmdopts_t *cmdopts)
 
 int addopt(char *optstr, int maxlen, char *s)
 {
-	int n;
-	int m;
+	size_t n;
+	size_t m;
 
 	n = strlen(optstr);
 	m = n + strlen(s) + 1;
-	if (m > maxlen) {
+	if (m > (size_t) maxlen) {
 		return 1;
 	}
 	if (n > 0) {
